requirements.txt
kafka-python==2.0.2
snowflake-snowpark-python==1.47.0
python-dotenv==1.0.1